ddehun joon plus

+++ pintos/src/vm/swap.h
@@ -5,7 +5,7 @@
 void swap_init(void);
 
 void swap_in (struct sup_pte * page, void * kpage);
-uint32_t swap_out(void * kpage );
+size_t swap_out(void * kpage );
 void swap_set(uint32_t swap_idx, bool value); 
 #endif
